HPCScheduler instance knows which HPCProcesses it is responsible for and cancels them safely
cancelation appeared to being executing on multiple threads and causing data-races.
This is not a performance critical operation, and therefore we use a high-level lock
to ensure that all processes are canceled in a single thread.

shutdown calls cancel to ensure that all owned processes are cleaned up.

signal handling is elevated to the submit_and_wait function and unsures shutdown is
called by only one thread at a time